+++ src/main.rs
@@ -17,7 +17,7 @@ use scene::Scene;
 use v3::{P3, V3};
 
 pub const BG_COLOR: Color = Color::new(0.7, 0.8, 1.0); // default scene background color
-pub const ASPECT_RATIO: f64 = 16.0 / 10.0; // image aspect ratio
+pub const ASPECT_RATIO: f32 = 16.0 / 10.0; // image aspect ratio
 pub const IMAGE_WIDTH: u16 = 1000; // image width in pixels
 pub const SAMPLES_PER_PIXEL: u16 = 4500; // number of random samples per pixel
 pub const STEP_SIZE: u16 = 100; // number of samples per render step
@@ -28,14 +28,14 @@ pub const SCENE_PATH: &str = "scene.toml";
 #[macro_export]
 macro_rules! p {
     ($x:expr, $y:expr, $z:expr) => {
-        P3::new($x as f64, $y as f64, $z as f64)
+        P3::new($x as f32, $y as f32, $z as f32)
     };
 }
 
 #[macro_export]
 macro_rules! v {
     ($x:expr, $y:expr, $z:expr) => {
-        V3::new($x as f64, $y as f64, $z as f64)
+        V3::new($x as f32, $y as f32, $z as f32)
     };
 }
